+++ src/fruits/mod.rs
@@ -457,6 +457,12 @@ impl FruitsPerformanceAttributes {
     pub fn pp(&self) -> f64 {
         self.pp
     }
+
+    /// Return the maximum combo of the map.
+    #[inline]
+    pub fn max_combo(&self) -> usize {
+        self.attributes.max_combo
+    }
 }
 
 impl From<FruitsPerformanceAttributes> for FruitsDifficultyAttributes {