+++ src/taiko/mod.rs
@@ -259,6 +259,12 @@ impl TaikoPerformanceAttributes {
     pub fn pp(&self) -> f64 {
         self.pp
     }
+
+    /// Return the maximum combo of the map.
+    #[inline]
+    pub fn max_combo(&self) -> usize {
+        self.attributes.max_combo
+    }
 }
 
 impl From<TaikoPerformanceAttributes> for TaikoDifficultyAttributes {